Refactored state to save only in background script
Having local storage being used directly from State was creating issues
in Firefox. In some instances the state woul be reset to an empy object
possibly due to errors related with multiple listeners being added to
the local storage and fired asynchronously. Now the state is not
considered a dao and is only saved from he background script which
avoids any of the verified problems.